Update to 0.2.6. See CHANGES.md.

+++ lib/common/common_actions_1.sh
@@ -1,13 +1,12 @@
-msg_print note "Copying file into rootfs..."
-cp -arf ./lib/distr/$distr/rootfs/* $dir/
-cp -af $ALEXPRO100_LIB_LOCATION $dir/root/
+msg_print note "Copying files into rootfs..."
+[[ -d ./lib/distr/$distr/rootfs ]] && cp -arf ./lib/distr/$distr/rootfs/* $dir/
 
 msg_print note "Saving configuration..."
 [[ -z $LANG ]] && LANG="en_US.UTF-8"
-echo -e "LANG=$LANG " >> $dir/root/configuration
-for var in ${var_list[@]}; do
-  echo -ne "$var=\"${!var}\" " >> $dir/root/configuration
-done
+[[ ! -d $dir/root ]] && mkdir $dir/root
+cp -af $ALEXPRO100_LIB_LOCATION $dir/root/
+echo -e "LANG=$LANG" >> $dir/root/configuration
+cat $profile_path >> $dir/root/configuration
 
 msg_print note "Configuring hosts..."
 echo "127.0.0.1	localhost
@@ -25,7 +24,7 @@ fi
 
 if [[ $fstab == "1" ]]; then
   msg_print note "Generationg fstab..."
-  mv $dir/etc/fstab{,.bak}
+  [[ -f $dir/etc/fstab ]] && mv $dir/etc/fstab{,.bak}
   echo '# Static information about the filesystems.
 # See fstab(5) for details.
 
@@ -34,3 +33,18 @@ if [[ $fstab == "1" ]]; then
 ' >> $dir/etc/fstab
   genfstab_light $dir >> $dir/etc/fstab
 fi
+
+msg_print note "Creating config script..."
+echo -ne "#!/bin/bash\n\nset -e; cd /root/\n" >> $dir/root/pi_s1.sh
+cat ./lib/common/rootfs_scripts/lib_connect.sh >> $dir/root/pi_s1.sh
+echo "
+echo 'Getting configuration...'
+if [[ -z /root/configuration ]]; then
+  return_err 'No configuration file!'
+else
+  source /root/configuration
+fi
+" >> $dir/root/pi_s1.sh
+cat ./lib/common/rootfs_scripts/base_change.sh >> $dir/root/pi_s1.sh
+
+msg_print note "Starting distr step..."
